backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.2.25
click==8.3.1
construct==2.10.70
construct-typing==0.7.0
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonalias==0.1.1
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
slowapi==0.1.9
tenacity==9.1.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
solana==0.36.11
solders==0.27.1
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.41.0
websockets==15.0.1
